Changing code specifically to fix api-extractor bugs
