+++ src/components/AnimatePresence/index.tsx
@@ -1,3 +1,4 @@
+import * as React from "react"
 import {
     useContext,
     useRef,
@@ -8,12 +9,15 @@ import {
     ReactNode,
     FunctionComponent,
 } from "react"
-import * as React from "react"
 import { AnimatePresenceProps } from "./types"
 import { SharedLayoutContext } from "../AnimateSharedLayout/SharedLayoutContext"
-import { SharedLayoutTree, SharedBatchTree } from "../AnimateSharedLayout/types"
 import { useForceUpdate } from "../../utils/use-force-update"
 import { PresenceChild } from "./PresenceChild"
+import {
+    SharedLayoutTree,
+    SharedBatchTree,
+} from "../../components/AnimateSharedLayout/types"
+
 type ComponentKey = string | number
 
 function getChildKey(child: ReactElement<any>): ComponentKey {
@@ -127,10 +131,10 @@ export const AnimatePresence: FunctionComponent<AnimatePresenceProps> = ({
     // We want to force a re-render once all exiting animations have finished. We
     // either use a local forceRender function, or one from a parent context if it exists.
     let forceRender = useForceUpdate()
-    const magicContext = useContext(SharedLayoutContext)
+    const layoutContext = useContext(SharedLayoutContext)
 
-    if (isControlledSharedLayoutContext(magicContext)) {
-        forceRender = magicContext.forceRender
+    if (isControlledSharedLayoutContext(layoutContext)) {
+        forceRender = layoutContext.forceRender
     }
 
     const isInitialRender = useRef(true)
