+++ src/index.ts
@@ -1,24 +1,13 @@
-/**
- * Components
- */
+export { motion, useExternalRef, createMotionComponent } from "./motion"
 export {
-    motion,
-    useExternalRef,
-    createMotionComponent,
     HTMLMotionProps,
     SVGMotionProps,
     SVGAttributesAsMotionValues,
     ForwardRefComponent,
-} from "./motion"
-export { AnimatePresence } from "./components/AnimatePresence"
-export { AnimateSharedLayout } from "./components/AnimateSharedLayout"
-
-/**
- * Motion values
- */
+} from "./motion/types"
+export { useMotionValue } from "./value/use-motion-value"
 export { MotionValue, motionValue, PassiveEffect, Subscriber } from "./value"
 export { resolveMotionValue } from "./value/utils/resolve-motion-value"
-export { useMotionValue } from "./value/use-motion-value"
 export { useInvertedScale } from "./value/use-inverted-scale"
 export { useViewportScroll } from "./value/use-viewport-scroll"
 export { useTransform } from "./value/use-transform"
