+++ src/motion/index.tsx
@@ -1,45 +1,38 @@
+import * as React from "react"
 import { createMotionComponent } from "./component"
 import { createDomMotionConfig } from "./features/dom"
-import * as React from "react"
-import {
-    ReactHTML,
-    DetailedHTMLFactory,
-    HTMLAttributes,
-    PropsWithoutRef,
-    RefAttributes,
-    SVGAttributes,
-    ForwardRefExoticComponent,
-} from "react"
-import { HTMLElements, htmlElements } from "./utils/supported-elements"
-import { svgElements, SVGElements } from "./utils/supported-elements"
-import { MotionProps, MakeMotion } from "./types"
+import { HTMLMotionComponents, SVGMotionComponents } from "./types"
 
+/**
+ * These re-exports are to
+ */
 export { MotionContext } from "./context/MotionContext"
 export { MotionValuesMap } from "./utils/use-motion-values"
 export { useExternalRef } from "./utils/use-external-ref"
 export { ValueAnimationControls } from "../animation/ValueAnimationControls"
 export { createMotionComponent }
 
-export const htmlMotionComponents: HTMLMotionComponents = htmlElements.reduce(
-    (acc, Component) => {
-        const config = createDomMotionConfig(Component)
-        // Suppress "Expression produces a union type that is too complex to represent" error
-        // @ts-ignore
-        acc[Component] = createMotionComponent(config)
-        return acc
-    },
-    {} as HTMLMotionComponents
-)
+type Motion = HTMLMotionComponents & SVGMotionComponents
 
-export const svgMotionComponents: SVGMotionComponents = svgElements.reduce(
-    (acc, Component) => {
-        // Suppress "Expression produces a union type that is too complex to represent" error
-        // @ts-ignore
-        acc[Component] = createMotionComponent(createDomMotionConfig(Component))
-        return acc
-    },
-    {} as SVGMotionComponents
-)
+/**
+ * Convert any React component into a `motion` component. The provided component
+ * **must** use `React.forwardRef` to the underlying DOM component you want to animate.
+ *
+ * ```jsx
+ * const Component = React.forwardRef((props, ref) => {
+ *   return <div ref={ref} />
+ * })
+ *
+ * const MotionComponent = motion.custom(Component)
+ * ```
+ *
+ * @public
+ */
+function custom<Props>(Component: string | React.ComponentType<Props>) {
+    return createMotionComponent<Props>(createDomMotionConfig(Component))
+}
+
+const componentCache = new Map<string, any>()
 
 /**
  * HTML & SVG components, optimised for use with gestures and animation. These can be used as
@@ -60,103 +53,20 @@ export const svgMotionComponents: SVGMotionComponents = svgElements.reduce(
  *
  * @public
  */
-export const motion = {
-    /**
-     * Convert a custom React component into a `motion` component.
-     *
-     * It can also accept a string, to create [custom DOM elements](https://developer.mozilla.org/en-US/docs/Web/Web_Components/Using_custom_elements).
-     *
-     * ```jsx
-     * const Component = React.forwardRef((props: Props, ref) => {
-     *   return <div ref={ref} />
-     * })
-     *
-     * const MotionComponent = motion.custom<Props>(Component)
-     * ```
-     *
-     * @param Component -
-     */
-    custom: function custom<Props>(
-        Component: string | React.ComponentType<Props>
-    ) {
-        return createMotionComponent<Props>(createDomMotionConfig(Component))
-    },
-    ...htmlMotionComponents,
-    ...svgMotionComponents,
-}
+export const motion = new Proxy(
+    { custom },
+    {
+        get: (target: any, key: string) => {
+            if (key === "custom") return target.custom
 
-type UnwrapFactoryAttributes<F> = F extends DetailedHTMLFactory<infer P, any>
-    ? P
-    : never
-type UnwrapFactoryElement<F> = F extends DetailedHTMLFactory<any, infer P>
-    ? P
-    : never
-type UnwrapSVGFactoryElement<F> = F extends React.SVGProps<infer P> ? P : never
+            if (!componentCache.has(key)) {
+                componentCache.set(
+                    key,
+                    createMotionComponent(createDomMotionConfig(key))
+                )
+            }
 
-type HTMLAttributesWithoutMotionProps<
-    Attributes extends HTMLAttributes<Element>,
-    Element extends HTMLElement
-> = { [K in Exclude<keyof Attributes, keyof MotionProps>]?: Attributes[K] }
-
-/**
- * @public
- */
-export type HTMLMotionProps<
-    TagName extends keyof ReactHTML
-> = HTMLAttributesWithoutMotionProps<
-    UnwrapFactoryAttributes<ReactHTML[TagName]>,
-    UnwrapFactoryElement<ReactHTML[TagName]>
-> &
-    MotionProps
-
-/**
- * Motion-optimised versions of React's HTML components.
- *
- * @public
- */
-export type HTMLMotionComponents = {
-    [K in HTMLElements]: ForwardRefComponent<
-        UnwrapFactoryElement<ReactHTML[K]>,
-        HTMLMotionProps<K>
-    >
-}
-
-interface SVGAttributesWithoutMotionProps<T>
-    extends Pick<
-        SVGAttributes<T>,
-        Exclude<keyof SVGAttributes<T>, keyof MotionProps>
-    > {}
-
-/**
- * Blanket-accept any SVG attribute as a `MotionValue`
- * @public
- */
-export type SVGAttributesAsMotionValues<T> = MakeMotion<
-    SVGAttributesWithoutMotionProps<T>
->
-
-/**
- * @public
- */
-export interface SVGMotionProps<T>
-    extends SVGAttributesAsMotionValues<T>,
-        MotionProps {}
-
-/**
- * @public
- */
-export type ForwardRefComponent<T, P> = ForwardRefExoticComponent<
-    PropsWithoutRef<P> & RefAttributes<T>
->
-
-/**
- * Motion-optimised versions of React's SVG components.
- *
- * @public
- */
-export type SVGMotionComponents = {
-    [K in SVGElements]: ForwardRefComponent<
-        UnwrapSVGFactoryElement<JSX.IntrinsicElements[K]>,
-        SVGMotionProps<UnwrapSVGFactoryElement<JSX.IntrinsicElements[K]>>
-    >
-}
+            return componentCache.get(key)
+        },
+    }
+) as Motion
